Fix puppet taking ages on hosts with large AEM repository
During its run the AEM module searches the installation folder for a
already present crx-quickstart jar.

On hosts with a grown AEM instance the 'crx-quickstart/repository'
folder can become quite huge, causing the find call to take ages when
searching withing this folder.

To fix this issue the find call was modified to skip the folder
'crx-quickstart/repository'.

Fixes [#124]